refactor: 응집도에 따른 폴더 아키텍쳐 변경
