+++ features/auth/signup/hooks/use-signup.mutation.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 
-import { createSignup } from './fetchers';
+import { createSignup } from '../create-signup.api';
 
 export const useSignup = () => {
   return useMutation({
