+++ features/auth/signup/types.ts
@@ -1,4 +1,4 @@
-import type { SignupFormSchema } from './schemas';
+import type { SignupFormSchema } from './signup-form.schema';
 import type { z } from 'zod';
 
 export type SignupFormData = z.infer<typeof SignupFormSchema>;
